+++ tests/test_app.py
@@ -23,6 +23,7 @@ import pytest
 import re
 from threading import Timer
 
+import inmanta.util
 import utils
 from inmanta import const
 import signal
@@ -37,7 +38,7 @@ def get_command(tmp_dir, stdout_log_level=None, log_file=None, log_level_log_fil
         os.mkdir(directory)
     config_file = os.path.join(root_dir, "inmanta.cfg")
 
-    port = utils.get_free_tcp_port()
+    port = inmanta.util.get_free_tcp_port()
 
     with open(config_file, 'w+') as f:
         f.write("[config]\n")
